using 10MHz clock

+++ include/fifo.h
@@ -4,7 +4,7 @@
 #include "board.h"
 
 typedef struct Sweep{
-    uint16_t        sweepDuration; 
+    uint32_t        sweepDuration; 
     uint16_t        lighthouse; 
     uint16_t        id; 
     bool            vertical;
